+++ test/.gitignore
@@ -1,1 +1,1 @@
-/generated/**
+**/generated/
