+++ test/generator_test/golden/l10n.dart
