+++ test/generator_test/golden/l10n_en.dart
